Execute the main cohort evaluation

Execute the main cohort evaluation

execute(
  connectionDetails,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  oracleTempSchema = cohortDatabaseSchema,
  outputFolder,
  databaseId = "Unknown",
  databaseName = "Unknown",
  databaseDescription = "Unknown",
  cohortGroups = getCohortGroups(),
  createCohorts = TRUE,
  runCohortDiagnostics = TRUE,
  minCellCount = 5
)

Arguments

connectionDetails
An object of type connectionDetails as created using the createConnectionDetails function in the DatabaseConnector package.

cdmDatabaseSchema
Schema name where your patient-level data in OMOP CDM format resides. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortDatabaseSchema
Schema name where intermediate data can be stored. You will need to have write priviliges in this schema. Note that for SQL Server, this should include both the database and schema name, for example 'cdm_data.dbo'.

cohortTable
The name of the table that will be created in the work database schema. This table will hold the exposure and outcome cohorts used in this study.

oracleTempSchema
Should be used in Oracle to specify a schema where the user has write priviliges for storing temporary tables.

outputFolder
Name of local folder to place results; make sure to use forward slashes (/). Do not use a folder on a network drive since this greatly impacts performance.

databaseId
A short string for identifying the database (e.g. 'Synpuf').

databaseName
The full name of the database (e.g. 'Medicare Claims Synthetic Public Use Files (SynPUFs)').

databaseDescription
A short description (several sentences) of the database.

cohortGroups
A character vector indicating which cohort groups to evaluate. See getCohortGroups() for options.

createCohorts
Create the cohortTable table with the exposure and outcome cohorts?

runCohortDiagnostics
Run the various cohort diagnostics?

minCellCount
The minimum number of subjects contributing to a count before it can be included in packaged results.

Details

This function executes the cohort diagnostics.